+++ demo/App.js
@@ -15,6 +15,12 @@ import {
 
 const SelectableFoo = createSeletable(Foo);
 
+const elements = Array(14).fill(0).map((_, index) => (
+  {
+    id: index,
+  }
+));
+
 
 class App extends React.PureComponent {
   constructor(props) {
@@ -34,10 +40,14 @@ class App extends React.PureComponent {
   }
 
   render() {
+    const {
+      selectedList,
+    } = this.state;
     return (
       <Container>
         <SelectableGroup
-          selectedList={this.state.selectedList}
+          selectedList={selectedList}
+          uidList={elements.map(e => e.id)}
           onChange={this._handleChange}
           style={{
             display: 'flex',
@@ -46,8 +56,12 @@ class App extends React.PureComponent {
           }}
         >
           {
-            Array(14).fill(0).map((ele, index) => index).map(ele => (
-              <SelectableFoo uid={ele} key={ele} />
+            elements.map(ele => (
+              <SelectableFoo
+                key={ele.id}
+                uid={ele.id}
+                selected={selectedList.indexOf(ele.id) >= 0}
+              />
             ))
           }
           <NonSelectable />