+++ src/SelectableGroup.js
@@ -1,17 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import createSeletable from './createSeletable';
-
 import {
   listCompare,
+  removeElement,
 } from './utils';
+
+import {
+  ACTION_NAME,
+} from './constants';
 
 class SelectableGroup extends React.Component {
   static propTypes = {
     children: PropTypes.node,
     onChange: PropTypes.func,
     selectedList: PropTypes.array,
+    uidList: PropTypes.array,
     Component: PropTypes.oneOfType([
       PropTypes.string,
       PropTypes.func,
@@ -25,31 +29,17 @@ class SelectableGroup extends React.Component {
   }
 
   static childContextTypes = {
-    actions: PropTypes.shape({
+    [ACTION_NAME]: PropTypes.shape({
       oneClick: PropTypes.func,
       toggleClick: PropTypes.func,
       rangeSelect: PropTypes.func,
     }),
   }
 
-  constructor(props) {
-    super(props);
-
-    this._toggleClick = this._toggleClick.bind(this);
-    this._oneClick = this._oneClick.bind(this);
-    this._rangeSelect = this._rangeSelect.bind(this);
-
-    this.state = {
-      uidList: React.Children.map(props.children, child => (
-        child.props.uid
-      )),
-    };
-  }
-
   getChildContext() {
     return (
       {
-        actions: {
+        [ACTION_NAME]: {
           oneClick: this._oneClick,
           toggleClick: this._toggleClick,
           rangeSelect: this._rangeSelect,
@@ -58,85 +48,64 @@ class SelectableGroup extends React.Component {
     );
   }
 
-  componentWillReceiveProps(nextProps) {
-    const uidList = React.Children.map(nextProps.children, child => (
-      child.props.uid
-    ));
+  _toggleClick = (uid) => {
+    const {
+      selectedList: prevSelectedList,
+      onChange,
+    } = this.props;
 
-    this.setState({
+    const index = prevSelectedList.indexOf(uid);
+
+    const selectedList = index < 0 ?
+      prevSelectedList.concat([uid]) :
+      removeElement(index)(prevSelectedList);
+
+    onChange(selectedList);
+  }
+
+  _oneClick = (uid) => {
+    const {
+      selectedList: prevSelectedList,
+      onChange,
+    } = this.props;
+
+    const index = prevSelectedList.indexOf(uid);
+
+    const isSelected = prevSelectedList.length > 1 || index < 0;
+    const selectedList = isSelected ? [uid] : removeElement(index)(prevSelectedList);
+
+    return onChange(selectedList);
+  }
+
+  _rangeSelect = (uid) => {
+    const {
+      selectedList,
       uidList,
-    });
-  }
+      onChange,
+    } = this.props;
+    if (selectedList.length === 0) return null;
 
-  _toggleClick(uid) {
-    const selectedList = [...this.props.selectedList];
+    const allCompare = listCompare(uidList);
 
-    const index = selectedList.indexOf(uid);
-
-    if (index < 0) {
-      selectedList.push(uid);
-    } else {
-      selectedList.splice(index, 1);
-    }
-
-    this.props.onChange(selectedList);
-  }
-
-  _oneClick(uid) {
-    let selectedList = [...this.props.selectedList];
-
-    const index = selectedList.indexOf(uid);
-
-    // more one element were selected
-    if (selectedList.length > 1) {
-      selectedList = [uid];
-    } else if (index < 0) {
-      selectedList = [uid];
-    } else {
-      selectedList.splice(index, 1);
-    }
-
-    this.props.onChange(selectedList);
-  }
-
-  _rangeSelect(uid) {
-    const selectedList = [...this.props.selectedList];
-    if (selectedList.length === 0) return;
-
-    const allCompare = listCompare(this.state.uidList);
-
-    this.props.onChange(allCompare(selectedList, uid));
+    return onChange(allCompare(selectedList, uid));
   }
 
 
   render() {
-    const { Component, children, selectedList, ...rest } = this.props;
-
-    const newChildren = React.Children.map(children, (child) => {
-      // check if child have selected in props
-      if (typeof child.type === 'function' && child.type.name === createSeletable().name) {
-        // clone a props and replace selected for new selectedList
-        const props = {
-          ...child.props,
-          selected: (selectedList.indexOf(child.props.uid) >= 0),
-        };
-
-        // clone a children
-        const { children: c } = child;
-
-        // clone a child
-        return React.createElement(child.type, props, c);
-      }
-
-      // if the child is not selectable, don't do anything
-      return child;
-    });
+    const {
+      Component,
+      children,
+      selectedList,
+      uidList,
+      onChange,
+      ...rest
+    } = this.props;
 
     return (
       <Component
         {...rest}
       >
-        {newChildren}
+        {children}
       </Component>
     );
   }