+++ src/createSeletable.js
@@ -1,11 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import {
+  ACTION_NAME,
+} from './constants';
+
 // High Order Component
 const createSeletable = (WrappedComponent) => {
   class SelectableComponent extends React.PureComponent {
     static contextTypes = {
-      actions: PropTypes.shape({
+      [ACTION_NAME]: PropTypes.shape({
         oneClick: PropTypes.func,
         toggleClick: PropTypes.func,
         rangeSelect: PropTypes.func,
@@ -21,18 +25,8 @@ const createSeletable = (WrappedComponent) => {
       onClick: PropTypes.func,
     }
 
-    static defaultProps = {
-      selected: false,
-    }
-
-    constructor(props) {
-      super(props);
-
-      this._handleClick = this._handleClick.bind(this);
-    }
-
-    _handleClick(e) {
-      const actions = this.context.actions;
+    _handleClick = (e) => {
+      const actions = this.context[ACTION_NAME];
 
       e.preventDefault();