+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## [Unreleased]
+## Unreleased
+* uidList is given via props
+* inject `selected` via props by developer
 
-## [1.0.0]
+## 1.0.0
 * context
-* use `React.Children.map` to inject selected property
+* use `React.Children.map` to inject `selected` property